+++ UpDockLicenseManager/Services/HealthResponse.swift
@@ -8,14 +8,18 @@
 import Foundation
 
 struct HealthResponse: Codable {
-    
+
     let status: String
-    
+
     let php: String
-    
+
     let time: String
-    
+
     let transactionsWritable: Bool
-    
+
     let fulfilledWritable: Bool
+
+    let licensesWritable: Bool?
+
+    let activationsWritable: Bool?
 }